#143 Remove interface, Change the method naming
